matching: bank 80 repaired drafts and lift 80043300 from asm to C
A draft-bank pass over the regenerated m2c sources: 80 functions that
differed from retail by a handful of instruction words were repaired to
byte matches (9,460 B, all complete C functions), and tools/repair_drafts.py
plus the idiom catalogue in docs/MATCHING.md record the mechanical edits
that turned them (field widths, indexed bases, commuted operands, retry
back-edges, delay-slot sinks, and the rest).

func_80043300 is the load-bearing one for the port: it was an instruction
__asm__ overlay, and it is now source C that rebuilds 38/38 words. That
moves 152 B of the startup dispatcher from assembly (word-exact but
unportable) into the native-executable column. Two further near-miss
drafts are banked in-tree (func_8003E248 31->36/39, func_80153D7C
27->28/33); a fourth attempt on func_80154ED8 regressed a 48/49 draft and
was dropped.

+++ src/main/8002b064.c
@@ -27,7 +27,7 @@ s32 func_8002B064(u8 *arg0) {
     s32 temp_a1;
     s32 var_v1;
     u8 *var_a0;
-    u8 temp_v0;
+    s32 temp_v0;
 
     var_a0 = arg0;
     var_v1 = 0;
